put region history in its own file
